Added Favorites page, indexedDB persist, modified NavBar, spaceXLaunches page, and rocketsDetails
Created the Favorites page to display all of the rockets that were selected as favorites (still unfinished).  Modified the NavBar to be responsive when display size is changed. Modified rocketsDetails design. Added persistence for favorites store to indexedDB. Created a Nuxt middleware that redirects users to launch page if page does not exist.